Fixed links and added detail.
After the repo restructure some of the URLs were 404ed. I fixed those, and while I was at it, added some additional functionality. 

I just realized that if you mix the main NAPP readme and this readme by combining
1. `helm install open5gs openverso/open5gs --values 5gSA_ues_values.yaml` which puts 10 UEs in the MongoDB
2. 'helm install ueransim-ues-first-batch openverso/ueransim-ues --values 0_450_ue_values.yaml` which tries to establish connections for 450 UEs

that only 10 get connections. The rest fail authentication. 

Since users will want to, as I wanted to, skip the first part since it was done in the main readme, we sould either clarify here or, better yet, move the procedure here to a complete description in the main readme.